+++ tendermint/src/block/height.rs
@@ -66,6 +66,18 @@ impl From<u8> for Height {
     }
 }
 
+impl PartialEq<u64> for Height {
+    fn eq(&self, other: &u64) -> bool {
+        self.value() == *other
+    }
+}
+
+impl PartialEq<u32> for Height {
+    fn eq(&self, other: &u32) -> bool {
+        self.value() == *other as u64
+    }
+}
+
 impl Height {
     /// Get inner integer value. Alternative to `.0` or `.into()`
     pub fn value(&self) -> u64 {
@@ -141,5 +153,11 @@ mod tests {
             Height::try_from(2_u64).unwrap().value(),
             Height::from(2_u32).value()
         );
+    }
+
+    #[test]
+    fn equivalence_check() {
+        assert!(Height::try_from(10_u64).unwrap() == 10u64);
+        assert!(Height::try_from(10_u64).unwrap() == 10u32);
     }
 }